fix pointer truncation in CToPascalCall, and some wrong ctop_flags

+++ src/include/SoundMgr.h
@@ -209,7 +209,7 @@ typedef void *AudioSelectionPtr;
 typedef void *SPBPtr;
 #endif
 
-typedef unsigned long UnsignedFixed;
+typedef uint32_t UnsignedFixed;
 typedef Ptr CompressionInfoPtr;
 
 extern trap void C_SndGetSysBeepState(GUEST<INTEGER> *statep);
